Take both read and written bytes in account for throughput
We were previously reporting memory throughput without considering
whether the measured routine actually does both reads and writes. For
the purpose of comparing different implementations of the same routines,
this is not a big issue, but for meaningful comparing of e.g. write
performance and copy performance, it is good to get the right numbers.